PythonBackend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0
demucs>=4.0.0
torch>=2.0
openai-whisper>=20231117
yt-dlp>=2024.1.0
librosa>=0.10.0
numpy>=1.24.0
